refactor: [AICPS-171] apply review feedback in TimeUpPopup
