+++ src/pages/AptitudeTest/components/Testing.tsx
@@ -15,7 +15,7 @@ import { useDeleteExam, useGetExamSet, useSubmitDraftQuestion, useSubmitExam } f
 import { AnswerChoice, Question } from '@app/interface/examSet.interface';
 import './QuestionIndexPanel.scss';
 import { useLocation } from 'react-router-dom';
-import TimeUpPopup from './TimeUpPopupProps';
+import TimeUpPopup from './TimeUpPopup';
 
 const Testing = () => {
   const { state } = useLocation();
